+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Renamed `ESUUIDString()` to `+[NSString UUIDString]`.
 - Removed `ESIsStringWithAnyText()`, `ESIsArrayWithItems()`, `ESIsDictionaryWithItems()`, `ESIsSetWithItems()`, `ESIsOrderedSetWithItems()`, `ESCreateNonretainedMutableSet()`, `ESCreateNonretainedMutableArray()`, `ESCreateNonretainedMutableDictionary()`.
 - Removed named color methods: `+es_redNavigationBarColor`, `+es_primaryButtonColor`, `+es_twitterColor` etc. [094cc70](https://github.com/ElfSundae/ESFramework/commit/094cc70ab39b9b272d333287a2ef72eecb08952a)
+- Removed utilities methods from [BButton](https://github.com/jessesquires/BButton)
 - Improved `ESRandomStringOfLength()`.
 - Renamed files from "ESAdditions" to "ESExtension".